Seperate discovery from classification. Add LogSummary. Merge timestampsCoarsen and VariantFilter into one component: ModifyLog.
